pyproject.toml
[build-system]
requires = ["uv_build >= 0.9.5, < 0.10.0"]
build-backend = "uv_build"

[project]
name = "pynamuh"
version = "0.1.0dev1"
description = "Python wrapper for Namuh Securities OpenAPI"
readme = "README.md"
requires-python = ">=3.13, <4.0"
license = "MIT"
license-files = ["LICEN[CS]E*"]
authors = [
    {name = "odumag99", email = "odumag99@gmail.com"}
]
keywords = ["namuh", "openAPI", "stock", "trading", "NH", "WMCA", "investment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pydantic>=2.11.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "pywin32>=311",
]

[project.urls]
Homepage = "https://github.com/yourusername/pynamuh"
Documentation = "https://github.com/yourusername/pynamuh/blob/main/README.md"
Repository = "https://github.com/yourusername/pynamuh"
Issues = "https://github.com/yourusername/pynamuh/issues"

[tool.uv]
package = true

[tool.uv.sources]
pynamuh = { workspace = true, editable = true }

[tool.uv.workspace]
members = [
    ".",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pynamuh",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]
